Transaction Log Service Class
