+++ TM1py/Services/TM1Service.py
@@ -1,5 +1,6 @@
 import pickle
 
+
 from TM1py.Services import HierarchyService, SecurityService, ApplicationService, SubsetService, ServerService, \
     MonitoringService, ProcessService, PowerBiService, AnnotationService, ViewService, RestService, CellService, \
     ChoreService, DimensionService, CubeService, ElementService, SandboxService, GitService
@@ -8,7 +9,7 @@ from TM1py.Services.JobService import JobService
 from TM1py.Services.UserService import UserService
 from TM1py.Services.ThreadService import ThreadService
 from TM1py.Services.SessionService import SessionService
-
+from TM1py.Services.TransactionLogService import TransactionLogService
 
 class TM1Service:
     """ All features of TM1py are exposed through this service
@@ -77,6 +78,7 @@ class TM1Service:
         self.users = UserService(self._tm1_rest)
         self.threads = ThreadService(self._tm1_rest)
         self.sessions = SessionService(self._tm1_rest)
+        self.transaction_logs = TransactionLogService(self._tm1_rest)
 
     def logout(self, **kwargs):
         self._tm1_rest.logout(**kwargs)
