+++ TM1py/__init__.py
@@ -67,5 +67,6 @@ from TM1py.Services.JobService import JobService
 from TM1py.Services.UserService import UserService
 from TM1py.Services.ThreadService import ThreadService
 from TM1py.Services.SessionService import SessionService
+from TM1py.Services.TransactionLogService import TransactionLogService
 
 __version__ = "2.0"
